Adicionar método de salvar computadores
